Move ActiveOffersTable to OffersTabContent
