+++ src/pages/OffersPage/components/OffersTabContent/components/ActiveOffersTable/ActiveOffersTable.module.less
@@ -1,9 +1,5 @@
-.tableRoot {
-  flex-grow: 1;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  overflow: hidden;
+.tableWrapper {
+  max-height: 440px;
 }
 
 .tableRoot table {
