+++ src/pages/OffersPage/components/OffersTabContent/components/ActiveOffersTable/Summary.tsx
