+++ src/pages/OffersPage/components/OffersTabContent/components/ActiveOffersTable/TableCells/APRCell.tsx
