+++ src/pages/OffersPage/components/OffersTabContent/components/ActiveOffersTable/TableCells/index.ts
