+++ src/pages/OffersPage/components/OffersTabContent/components/ActiveOffersTable/helpers.ts
