+++ src/pages/OffersPage/components/OffersTabContent/components/ActiveOffersTable/hooks/index.ts
