+++ src/pages/OffersPage/components/OffersTabContent/components/ActiveOffersTable/hooks/useLenderLoansAndOffers.ts
