+++ src/pages/OffersPage/components/OffersTabContent/components/ActiveOffersTable/index.ts
