+++ src/pages/OffersPage/components/OffersTabContent/components/OfferCard/OfferCard.module.less
@@ -34,6 +34,7 @@
   display: flex;
   align-items: center;
   gap: 16px;
+  white-space: nowrap;
 }
 .collectionImage {
   width: 72px;
